+++ js/util.js
@@ -1,0 +1,7 @@
+const returnRandomInteger = function (min, max) {
+  min=Math.ceil(min);
+  max=Math.floor(max);
+  const rez = Math.floor(Math.random() * (max - min + 1)) + min;
+  return rez;
+};
+export {returnRandomInteger};
